# server-management-system/api-gateway/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Server Management System API
  description: |
    API quản lý 10.000 server.
    
    ## Tính năng chính
    - **Auth**: Đăng ký, đăng nhập, JWT, RBAC (Admin / Operator / Viewer)
    - **Server CRUD**: Tạo, xem (filter/sort/paging), cập nhật, xóa
    - **Import/Export Excel**: Upload .xlsx để import hàng loạt, export ra file
    - **Monitoring**: Health-check TCP 10.000 server mỗi 60s, ghi Elasticsearch
    - **Báo cáo**: Báo cáo định kỳ qua Email + API báo cáo chủ động
    
    ## Authentication
    Tất cả API (trừ register, login, refresh) yêu cầu JWT Bearer token.
    Mỗi API có scope riêng (xem từng endpoint).
    
    ## Error Format
    ```json
    {
      "status": "error",
      "code": 42201,
      "message": "Validation failed",
      "errors": [{ "field": "ipv4", "code": "INVALID_FORMAT", "message": "..." }],
      "meta": { "request_id": "req-uuid", "timestamp": "2026-06-10T10:00:00Z" }
    }
    ```
  version: 1.0.0
  contact:
    name: SMS Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ POST /auth/login.
        Token hết hạn sau 15 phút, dùng POST /auth/refresh để làm mới.

  # ═══════════════════════════════════════
  # Schemas
  # ═══════════════════════════════════════
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, email, password, full_name]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: admin01
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          minLength: 8
          example: "SecureP@ss123"
        full_name:
          type: string
          example: "Nguyen Van A"
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin01
        password:
          type: string
          example: "SecureP@ss123"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        username:
          type: string
          example: admin01
        email:
          type: string
          example: admin@example.com
        full_name:
          type: string
          example: "Nguyen Van A"
        role:
          type: string
          example: admin
        scopes:
          type: array
          items:
            type: string
          example: ["server:create", "server:read", "server:update", "server:delete", "server:import", "server:export", "report:view", "report:send", "user:manage"]
        is_active:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          example: "2026-06-10T08:00:00Z"
        created_at:
          type: string
          format: date-time

    # ── User Management ──
    UpdateUserRoleRequest:
      type: object
      required: [role_name]
      properties:
        role_name:
          type: string
          enum: [admin, operator, viewer]
          description: "Vai trò mới cho user"
          example: operator
    UserListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 3
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'

    # ── Server ──
...
    ExportRequest:
      type: object
      properties:
        filter:
          type: object
          properties:
            status:
              type: string
              enum: [on, off]
            server_name:
              type: string
            ipv4:
              type: string
            location:
              type: string
            os:
              type: string
        sort_by:
          type: string
          enum: [server_id, server_name, status, ipv4, location, created_at, updated_at]
          example: created_at
        sort_order:
          type: string
          enum: [asc, desc]
          example: desc

    # ── Report ──
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
        meta:
          $ref: '#/components/schemas/MetaInfo'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 42201
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        meta:
          $ref: '#/components/schemas/MetaInfo'
    FieldError:
      type: object
      properties:
        field:
          type: string
          example: ipv4
        code:
          type: string
          example: INVALID_FORMAT
        message:
          type: string
          example: "IPv4 must be in format x.x.x.x"
    MetaInfo:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440099"
        timestamp:
          type: string
          format: date-time
          example: "2026-06-10T10:00:00Z"

# ═══════════════════════════════════════
# Paths
# ═══════════════════════════════════════
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản mới
      operationId: register
      security: []  # Public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '400':
          description: Request body không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username hoặc email đã tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập, nhận JWT token
      operationId: login
      security: []  # Public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Sai username hoặc password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Làm mới access token bằng refresh token
      operationId: refreshToken
      security: []  # Public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token mới
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh token không hợp lệ hoặc hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Đăng xuất, đưa token vào blacklist
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Đăng xuất thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Logged out successfully"
        '401':
          description: Token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    get:
      tags: [Auth]
      summary: Xem thông tin profile của user hiện tại
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          description: Token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══ USERS ═══
  /auth/users:
    get:
      tags: [Users]
      summary: Lấy danh sách người dùng (Admin only)
      description: |
        **Scope:** `user:manage`
        **Role:** Admin only
      operationId: listUsers
      security:
        - BearerAuth: [user:manage]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Trang hiện tại
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Số lượng user mỗi trang
      responses:
        '200':
          description: Danh sách người dùng
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Thiếu JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Không có scope `user:manage`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/users/{user_id}/role:
    put:
      tags: [Users]
      summary: Thay đổi role của người dùng (Admin only)
      description: |
        Admin có thể nâng cấp hoặc hạ cấp role của user khác.
        Không thể tự thay đổi role của chính mình.
        **Scope:** `user:manage`
        **Role:** Admin only
      operationId: updateUserRole
      security:
        - BearerAuth: [user:manage]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: Role đã được cập nhật
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '400':
          description: Không thể tự đổi role của mình hoặc request không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Thiếu JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Không có scope `user:manage`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User không tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══ SERVERS ═══
...
  /servers/export:
    post:
      tags: [Import/Export]
      summary: Export danh sách server ra file Excel
      description: |
        Trả về file .xlsx. Có filter, sort tùy chọn.
        **Scope:** `server:export`
        **Role:** Admin only
      operationId: exportServers
      security:
        - BearerAuth: [server:export]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: File Excel (binary stream)
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: 'attachment; filename="servers_export_20260610.xlsx"'
              schema:
                type: string
        '401':
          description: Thiếu JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Không có scope `server:export`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══ REPORTS ═══
...